pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "hud-humanevalplus"
version = "0.1.0"
dependencies = [ "hud-python>=0.4.47", "datasets>=4.1.1", "pytest>=7.0.0", "pytest-json-report>=1.5.0",]
requires-python = ">=3.11,<3.13"
description = "HUD HumanEval+ Evaluation with Test Execution as MCP tools."
license = "MIT"
[[project.authors]]
name = "Martin Weyssow"
email = "weyssowm@gmail.com"

[tool.hud]
image = "humanevalplus:dev"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [ "src/hud_controller",]
